+++ tests/testthat/test_ebird-species.r
@@ -48,6 +48,11 @@ test_that("ebird_species works for species with non-ascii characters", {
 test_that("ebird_species handles versions correctly", {
   skip_on_cran()
   skip_if_offline()
+  # check that taxonomy api is accessible
+  e <- tryCatch(get_ebird_taxonomy(2017), error = function(e) e)
+  if (inherits(e, "error")) {
+    skip("eBird API inaccessible")
+  }
   expect_equal(ebird_species("Cordillera Azul Antbird"), "Myrmoderus eowilsoni")
   expect_equal(ebird_species("Cordillera Azul Antbird", 
                              taxonomy_version = 2017),