+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - new `auk_county()` filter
 - new `auk_year()` filter
+- Drop taxonomy warnings since there was no taxonomy update this year
 
 # auk 0.4.1